fix: invalid filter categories and difficulty
- created custom errors
- add logging for errors
