+++ question-service/errors/BaseError.js
@@ -1,0 +1,8 @@
+class BaseError extends Error {
+    constructor(statusCode, message) {
+        super(message);
+        this.statusCode = statusCode; 
+    }
+}
+
+export default BaseError;
