+++ question-service/middlewares/errorHandler.js
@@ -1,13 +1,19 @@
-
 const errorHandler = (err, req, res, next) => {
-    if (err.isJoi) {
-      // VALIDATION ERRORS  
-      return res.status(400).json({ message: err.details[0].message }); 
-    }
-    
-    console.error(err); 
-    
-    return res.status(500).json({ message: "Internal Server Error" });
-  };
-  
-  export default errorHandler;
+
+  const errStatus = err.statusCode || 500;
+  const errMsg = err.message || "Internal Server Error";
+
+  console.log({
+    success: false,
+    status: errStatus,
+    message: errMsg
+  })
+
+  res.status(errStatus).json({
+    success: false,
+    status: errStatus,
+    message: errMsg
+  });
+};
+
+export default errorHandler;
